+++ CHANGELOG.md
@@ -5,6 +5,12 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+* `handle_resource_usage` to handle the `ResourceUsage` request.
+
 ## [0.2.3] - 2022-05-17
 
 ### Added
@@ -51,6 +57,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 * `send_frame` and `recv_frame` to send and receive length-delimited frames.
 
+[Unreleasd]: https://github.com/petabi/oinq/compare/0.2.3...main
 [0.2.3]: https://github.com/petabi/oinq/compare/0.2.2...0.2.3
 [0.2.2]: https://github.com/petabi/oinq/compare/0.2.1...0.2.2
 [0.2.1]: https://github.com/petabi/oinq/compare/0.2.0...0.2.1